+++ components/mip-position-observer/service/position-observer-service.js
