+++ components/mip-position-observer/service/position-observer-worker.js
